+++ src/client/stores/dictionary.ts
@@ -1,11 +1,8 @@
 import delve from 'dlv'
 import merge from 'deepmerge'
 import { writable, derived } from 'svelte/store'
-
 import { Dictionary } from '../types/index'
-
-import { lookupMessage } from '../includes/lookup'
-
+import { lookup,lookupCache } from '../includes/lookup'
 import { getFallbackOf } from './locale'
 
 
@@ -46,14 +43,14 @@ export function nestingLoop(maindict:any,locale:string,dict:any = {}) {
   for(const key in dict) {
     if (typeof dict[key] == 'object') dict[key] = nestingLoop(maindict,locale,dict[key])
     else if (typeof dict[key] == 'string') {
-      let not_found:string[] = []
-      while(true) {
+      const not_found:string[] = []
+      for(;;) {
         let nested = Array.from(dict[key].matchAll(/\{\{(.*?)\}\}/g),(m:any)=>m[1])
         nested = nested.filter(n=>!not_found.includes(n))
         if (!nested || nested.constructor != Array || nested.length === 0) break
         nested.forEach( (k:string) =>{
           const K = `{{${k}}}`
-          const v = k!='' ? lookupMessage(k,locale,false) || K : ''
+          const v = k!='' ? lookup(k,locale,false) || K : ''
           if (v == K) return not_found.push(k)
           dict[key] = dict[key].replace( new RegExp(K),v )
         })
@@ -68,7 +65,8 @@ export function addMessages(locale: string, ...partials: Dictionary[]) {
     dictionary[locale] = merge.all<Dictionary>(
       [getLocaleDictionary(locale) || {}].concat(partials)
     )
-    d[locale]=nestingLoop(d[locale],locale)
+    dictionary[locale]=nestingLoop(dictionary[locale],locale)
+    lookupCache[locale] = {}
     return d
   })
 }